Fix `root_search_requests_total` metric
